gfc-showcase-production/backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-filter==23.5
django-cors-headers==4.3.1
python-decouple==3.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.1.0
drf-spectacular==0.27.0
django-celery-beat==2.5.0
redis==5.0.1
